+++ jablib/src/main/java/org/jabref/logic/database/DuplicateCheck.java
@@ -76,8 +76,8 @@ public class DuplicateCheck {
 
     private static boolean haveSameIdentifier(final BibEntry one, final BibEntry two) {
         return one.getFields().stream()
-                .filter(field -> field.getProperties().contains(FieldProperty.IDENTIFIER))
-                .anyMatch(field -> two.getField(field).map(content -> one.getField(field).orElseThrow().equals(content)).orElse(false));
+                  .filter(field -> field.getProperties().contains(FieldProperty.IDENTIFIER))
+                  .anyMatch(field -> two.getField(field).map(content -> one.getField(field).orElseThrow().equals(content)).orElse(false));
     }
 
     private static boolean haveDifferentEntryType(final BibEntry one, final BibEntry two) {